Move all http endpoints to same file
